InvitationControllerTests completed and some errors fixed

+++ code/jvm/service/src/main/kotlin/pt/isel/InvitationService.kt
@@ -111,11 +111,17 @@ class InvitationService(
             return@run success(createdInvitation)
         }
 
-    fun acceptChannelInvitation(invitationId: Int): Either<InvitationError, Channel> =
+    fun acceptChannelInvitation(invitationId: Int, userId: Int): Either<InvitationError, Channel> =
         trxManager.run {
+            userRepo.findById(userId) ?: return@run failure(InvitationError.UserNotFound)
+            if (userId < 0) return@run failure(InvitationError.NegativeIdentifier)
+
             val invitation: ChannelInvitation =
                 invitationRepo.findChannelInvitationById(invitationId)
                     ?: return@run failure(InvitationError.InvitationNotFound)
+            if(invitation.receiver.id != userId) return@run failure(InvitationError.Unauthorized)
+            if(invitation.timestamp.plusDays(1).isBefore(LocalDateTime.now())) return@run failure(InvitationError.InvitationExpired)
+
             if (invitation.isUsed) return@run failure(InvitationError.InvitationAlreadyUsed)
             val channel =
                 channelRepo.addUserToChannel(invitation.receiver, invitation.channel, invitation.role)
@@ -123,9 +129,16 @@ class InvitationService(
             return@run success(channel)
         }
 
-    fun declineChannelInvitation(invitationId: Int): Either<InvitationError, Boolean> =
+    fun declineChannelInvitation(invitationId: Int, userId: Int): Either<InvitationError, Boolean> =
         trxManager.run {
-            invitationRepo.findChannelInvitationById(invitationId) ?: return@run failure(InvitationError.InvitationNotFound)
+            val invitation = invitationRepo.findChannelInvitationById(invitationId) ?: return@run failure(InvitationError.InvitationNotFound)
+
+            userRepo.findById(userId) ?: return@run failure(InvitationError.UserNotFound)
+            if (userId < 0) return@run failure(InvitationError.NegativeIdentifier)
+
+            if(invitation.receiver.id != userId) return@run failure(InvitationError.Unauthorized)
+            if(invitation.timestamp.plusDays(1).isBefore(LocalDateTime.now())) return@run failure(InvitationError.InvitationExpired)
+
             val declined = invitationRepo.deleteChannelInvitationById(invitationId)
             return@run success(declined)
         }
